refactor(report): report 생성 및 처리 로직을 통합하여 report 처리 단순화
