+++ immilog-core/src/main/java/com/backend/immilog/report/application/usecase/ReportUseCase.java
@@ -8,34 +8,23 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
-public interface ReportUseCase {
-    void execute(
-            String targetUserId,
-            ReportCommand command
-    );
+@Slf4j
+@Service
+public class ReportUseCase {
+    private final ReportService reportService;
 
-    @Slf4j
-    @Service
-    class Reporter implements ReportUseCase {
-        private final ReportService reportService;
+    public ReportUseCase(ReportService reportService) {
+        this.reportService = reportService;
+    }
 
-        public Reporter(ReportService reportService) {
-            this.reportService = reportService;
-        }
-
-        @Async
-        @DistributedLock(key = "'targetId : '", identifier = "#p0.toString()", expireTime = 5)
-        @Override
-        public void execute(
-                String targetId,
-                ReportCommand command
-        ) {
-            var reporterUserId = command.reporterUserId();
-            var reason = command.reason();
-            var isOther = reason.equals(ReportReason.OTHER);
-            var description = isOther ? command.description() : reason.getDescription();
-            reportService.report(targetId, reporterUserId, reason, description);
-            log.info("User {} reported by {}", targetId, reporterUserId);
-        }
+    @Async
+    @DistributedLock(key = "'targetId : '", identifier = "#p0.toString()", expireTime = 5)
+    public void execute(String targetId, ReportCommand command) {
+        var reporterUserId = command.reporterUserId();
+        var reason = command.reason();
+        var isOther = reason.equals(ReportReason.OTHER);
+        var description = isOther ? command.description() : reason.getDescription();
+        reportService.report(targetId, reporterUserId, reason, description);
+        log.info("User {} reported by {}", targetId, reporterUserId);
     }
 }
